Support autodetection interface methods in windows
